==> Scripts/02_data_analysis/01_diversity/04b_jaccard_tables.R <==
#Jaccard Tables + Plots
#goal: visualize jaccard components and format values into a table

#SET-UP ####
#load libraries 
library(tidyr)
library(tidyverse)
library(here)
library(dplyr)
library(rempsyc)

#read data
data3 <- read.csv(here::here("Processed_data",
                             "datasets",
                            "diversity",
                            "diversity_indices_all.csv"),
                 head=TRUE)

#table w/ all sets ####
data3$set_number <- as.numeric(data3$set_number) #make set-number numeric instead of character
data3 <- data3 %>% arrange(set_number) #arrange numbers in increasing order 
data3$set_number <- as.character(data3$set_number) #change set-number back to character
colnames(data3) <- c('site','Jaccards','Jaccards Turnover', 'Jaccards Nestedness') #change column names 
data3 <- select(data3, c('site','Jaccards','Jaccards Turnover', 'Jaccards Nestedness')) #select specific columns to be included in table

#make table!
my_table <- nice_table(
  data3[1:16, ], 
  title = c("Table 2", "Diversity Indices Dissimilarities"), 
  note = c("The data was calculated from adapted methods from Baselga & Leprieur (2015)"))

my_table


#barplot of components ####
#Make bar plot showing proportion nestedness + proportion turnover  

data <- select(data3, c('site', 'Jaccards Turnover', 'Jaccards Nestedness'))
colnames(data) <- c('site','Turnover', 'Nestedness')
data_long <- gather(data, dissimilarity, measurement, c('Turnover'):c('Nestedness'), factor_key=TRUE)

#order in decreasing nestedness, increasing turnover 
data_long$site <- factor(data_long$site,levels = c("1", "4", "5", "11",'16','15','6','10','9','14','12','8','3','13','2','7'))

#plot
plot <- ggplot(data_long, aes(fill=dissimilarity, y=measurement, x=site)) + 
  geom_bar(position = 'fill', stat="identity") +
  scale_fill_manual(values=c("#00AFBB", "#132B43")) +
  xlab("site") + ylab("Proportion of Jaccards Dissimilarity") +
  theme_classic()
plot

#save
ggsave("./Outputs/diversity/jacccomponents_all.png", 
       plot = plot,
       width = 10, height = 5, units = "in")
